requirements.txt
asgiref==3.3.1
Django==3.0
django-cors-headers==3.2.1
django-debug-toolbar==2.1
django-filter==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
